Fplit out first/last tests, and add a test for the First replacement operators
Continuing to work on #68
